Add files via upload
This is a simple python script for downloading several images at once from google.

It uses the tkinter library for generating a simple user interface, and lets the user type in a search and number of desired results, which then will be collected using web scraping with the Selenium webdriver, named in batch and added to a new folder in a location of your choosing.

If you want to use this code yourself, there are a few things you'll need to rename, but I have added comments in the code to hopefully make it pretty straight-forward.

This was intended to help me learn more about GUIs for Python, as well as web scraping. I imagine it could be a useful tool if you quickly want to generate training data for machine learning, etc.